backend/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
sqlalchemy
langchain
langchain-groq
langchain-core
python-dotenv
requests